Add files via upload
The Forest Fire Prediction project aims to develop a predictive model that can forecast the likelihood of forest fires based on various environmental and geographical factors. This project is crucial for implementing preventive measures, optimizing resource allocation for firefighting, and minimizing the devastating impact of forest fires on ecosystems, human life, and property.